Improved connection tracking
This refactor is necessary because on macOS because some calls will hang if the device is no longer connected. Also the user might not have granted bluetooth access to the application yet, the application will now instruct the user the access is missing. Also the application will now attempt to restart scanning after the access has been granted, although note that this does not always work because of limitations in macOS in which case a restart is required.