Use the "windows" subsystem on windows
This is instead of hiding the console window using `ShowWindow(x, SW_HIDE)`. This no longer works on Windows 11.

This was useful on Windows 10 because it allowed the application to display log lines when ran from an existing terminal. As an alternative, set the environment variable LOG_FILE to a valid path.